Возможность объявлять криптопары, по каждой будут сохраняться данные

+++ src/main/java/dev/zykov/model/DepthResponse.java
@@ -30,13 +30,13 @@ public class DepthResponse {
     private String symbol;
 
     @JsonProperty("U")
-    private Integer firstUpdateId;
+    private Long firstUpdateId;
 
     @JsonProperty("u")
-    private Integer lastUpdateId;
+    private Long lastUpdateId;
 
     @JsonProperty("pu")
-    private Integer lastUpdateIdInLastStream;
+    private Long lastUpdateIdInLastStream;
 
     @JsonProperty("b")
     private List<List<BigDecimal>> bids;
